@@ -5,160 +5,160 @@ using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
-using BoardBarons.Api.Data;
+//using BoardBarons.Api.Data;
 using BoardBarons.Business.Entities;
 
 namespace BoardBarons.Api.Controllers
 {
     public class EventsController : Controller
     {
-        private readonly BoardBaronsApiContext _context;
+        //private readonly BoardBaronsApiContext _context;
 
-        public EventsController(BoardBaronsApiContext context)
-        {
-            _context = context;
-        }
+        //public EventsController(BoardBaronsApiContext context)
+        //{
+        //    _context = context;
+        //}
 
-        // GET: Events
-        public async Task<IActionResult> Index()
-        {
-              return _context.Event != null ? 
-                          View(await _context.Event.ToListAsync()) :
-                          Problem("Entity set 'BoardBaronsApiContext.Event'  is null.");
-        }
+        //// GET: Events
+        //public async Task<IActionResult> Index()
+        //{
+        //      return _context.Event != null ? 
+        //                  View(await _context.Event.ToListAsync()) :
+        //                  Problem("Entity set 'BoardBaronsApiContext.Event'  is null.");
+        //}
 
-        // GET: Events/Details/5
-        public async Task<IActionResult> Details(Guid? id)
-        {
-            if (id == null || _context.Event == null)
-            {
-                return NotFound();
-            }
+        //// GET: Events/Details/5
+        //public async Task<IActionResult> Details(Guid? id)
+        //{
+        //    if (id == null || _context.Event == null)
+        //    {
+        //        return NotFound();
+        //    }
 
-            var @event = await _context.Event
-                .FirstOrDefaultAsync(m => m.Id == id);
-            if (@event == null)
-            {
-                return NotFound();
-            }
+        //    var @event = await _context.Event
+        //        .FirstOrDefaultAsync(m => m.Id == id);
+        //    if (@event == null)
+        //    {
+        //        return NotFound();
+        //    }
 
-            return View(@event);
-        }
+        //    return View(@event);
+        //}
 
-        // GET: Events/Create
-        public IActionResult Create()
-        {
-            return View();
-        }
+        //// GET: Events/Create
+        //public IActionResult Create()
+        //{
+        //    return View();
+        //}
 
-        // POST: Events/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("Name,DateEvent,Created_At,Updated_At,Desativated_At,Group_Id,Media_Id,Id")] Event @event)
-        {
-            if (ModelState.IsValid)
-            {
-                @event.Id = Guid.NewGuid();
-                _context.Add(@event);
-                await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
-            }
-            return View(@event);
-        }
+        //// POST: Events/Create
+        //// To protect from overposting attacks, enable the specific properties you want to bind to.
+        //// For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
+        //[HttpPost]
+        //[ValidateAntiForgeryToken]
+        //public async Task<IActionResult> Create([Bind("Name,DateEvent,Created_At,Updated_At,Desativated_At,Group_Id,Media_Id,Id")] Event @event)
+        //{
+        //    if (ModelState.IsValid)
+        //    {
+        //        @event.Id = Guid.NewGuid();
+        //        _context.Add(@event);
+        //        await _context.SaveChangesAsync();
+        //        return RedirectToAction(nameof(Index));
+        //    }
+        //    return View(@event);
+        //}
 
-        // GET: Events/Edit/5
-        public async Task<IActionResult> Edit(Guid? id)
-        {
-            if (id == null || _context.Event == null)
-            {
-                return NotFound();
-            }
+        //// GET: Events/Edit/5
+        //public async Task<IActionResult> Edit(Guid? id)
+        //{
+        //    if (id == null || _context.Event == null)
+        //    {
+        //        return NotFound();
+        //    }
 
-            var @event = await _context.Event.FindAsync(id);
-            if (@event == null)
-            {
-                return NotFound();
-            }
-            return View(@event);
-        }
+        //    var @event = await _context.Event.FindAsync(id);
+        //    if (@event == null)
+        //    {
+        //        return NotFound();
+        //    }
+        //    return View(@event);
+        //}
 
-        // POST: Events/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(Guid id, [Bind("Name,DateEvent,Created_At,Updated_At,Desativated_At,Group_Id,Media_Id,Id")] Event @event)
-        {
-            if (id != @event.Id)
-            {
-                return NotFound();
-            }
+        //// POST: Events/Edit/5
+        //// To protect from overposting attacks, enable the specific properties you want to bind to.
+        //// For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
+        //[HttpPost]
+        //[ValidateAntiForgeryToken]
+        //public async Task<IActionResult> Edit(Guid id, [Bind("Name,DateEvent,Created_At,Updated_At,Desativated_At,Group_Id,Media_Id,Id")] Event @event)
+        //{
+        //    if (id != @event.Id)
+        //    {
+        //        return NotFound();
+        //    }
 
-            if (ModelState.IsValid)
-            {
-                try
-                {
-                    _context.Update(@event);
-                    await _context.SaveChangesAsync();
-                }
-                catch (DbUpdateConcurrencyException)
-                {
-                    if (!EventExists(@event.Id))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
-                }
-                return RedirectToAction(nameof(Index));
-            }
-            return View(@event);
-        }
+        //    if (ModelState.IsValid)
+        //    {
+        //        try
+        //        {
+        //            _context.Update(@event);
+        //            await _context.SaveChangesAsync();
+        //        }
+        //        catch (DbUpdateConcurrencyException)
+        //        {
+        //            if (!EventExists(@event.Id))
+        //            {
+        //                return NotFound();
+        //            }
+        //            else
+        //            {
+        //                throw;
+        //            }
+        //        }
+        //        return RedirectToAction(nameof(Index));
+        //    }
+        //    return View(@event);
+        //}
 
-        // GET: Events/Delete/5
-        public async Task<IActionResult> Delete(Guid? id)
-        {
-            if (id == null || _context.Event == null)
-            {
-                return NotFound();
-            }
+        //// GET: Events/Delete/5
+        //public async Task<IActionResult> Delete(Guid? id)
+        //{
+        //    if (id == null || _context.Event == null)
+        //    {
+        //        return NotFound();
+        //    }
 
-            var @event = await _context.Event
-                .FirstOrDefaultAsync(m => m.Id == id);
-            if (@event == null)
-            {
-                return NotFound();
-            }
+        //    var @event = await _context.Event
+        //        .FirstOrDefaultAsync(m => m.Id == id);
+        //    if (@event == null)
+        //    {
+        //        return NotFound();
+        //    }
 
-            return View(@event);
-        }
+        //    return View(@event);
+        //}
 
-        // POST: Events/Delete/5
-        [HttpPost, ActionName("Delete")]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(Guid id)
-        {
-            if (_context.Event == null)
-            {
-                return Problem("Entity set 'BoardBaronsApiContext.Event'  is null.");
-            }
-            var @event = await _context.Event.FindAsync(id);
-            if (@event != null)
-            {
-                _context.Event.Remove(@event);
-            }
+        //// POST: Events/Delete/5
+        //[HttpPost, ActionName("Delete")]
+        //[ValidateAntiForgeryToken]
+        //public async Task<IActionResult> DeleteConfirmed(Guid id)
+        //{
+        //    if (_context.Event == null)
+        //    {
+        //        return Problem("Entity set 'BoardBaronsApiContext.Event'  is null.");
+        //    }
+        //    var @event = await _context.Event.FindAsync(id);
+        //    if (@event != null)
+        //    {
+        //        _context.Event.Remove(@event);
+        //    }
             
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
-        }
+        //    await _context.SaveChangesAsync();
+        //    return RedirectToAction(nameof(Index));
+        //}
 
-        private bool EventExists(Guid id)
-        {
-          return (_context.Event?.Any(e => e.Id == id)).GetValueOrDefault();
-        }
+        //private bool EventExists(Guid id)
+        //{
+        //  return (_context.Event?.Any(e => e.Id == id)).GetValueOrDefault();
+        //}
     }
 }
