@@ -5,7 +5,7 @@ using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
-using BoardBarons.Api.Data;
+//using BoardBarons.Api.Data;
 using BoardBarons.Business.Entities;
 using Microsoft.AspNetCore.Authorization;
 
@@ -15,153 +15,153 @@ namespace BoardBarons.Api.Controllers
     [Route("api/[controller]")]
     public class RegionsController : Controller
     {
-        private readonly IRegionService _service;
+        //private readonly IRegionService _service;
 
-        public RegionsController(IRegionService service)
-        {
-            _service = service;
-        }
+        //public RegionsController(IRegionService service)
+        //{
+        //    _service = service;
+        //}
 
-        // GET: Regions
-        public async Task<IActionResult> Index()
-        {
-              return _context.Region != null ? 
-                          View(await _context.Region.ToListAsync()) :
-                          Problem("Entity set 'BoardBaronsApiContext.Region'  is null.");
-        }
+        //// GET: Regions
+        //public async Task<IActionResult> Index()
+        //{
+        //      return _context.Region != null ? 
+        //                  View(await _context.Region.ToListAsync()) :
+        //                  Problem("Entity set 'BoardBaronsApiContext.Region'  is null.");
+        //}
 
-        // GET: Regions/Details/5
-        public async Task<IActionResult> Details(Guid? id)
-        {
-            if (id == null || _context.Region == null)
-            {
-                return NotFound();
-            }
+        //// GET: Regions/Details/5
+        //public async Task<IActionResult> Details(Guid? id)
+        //{
+        //    if (id == null || _context.Region == null)
+        //    {
+        //        return NotFound();
+        //    }
 
-            var region = await _context.Region
-                .FirstOrDefaultAsync(m => m.Id == id);
-            if (region == null)
-            {
-                return NotFound();
-            }
+        //    var region = await _context.Region
+        //        .FirstOrDefaultAsync(m => m.Id == id);
+        //    if (region == null)
+        //    {
+        //        return NotFound();
+        //    }
 
-            return View(region);
-        }
+        //    return View(region);
+        //}
 
-        // GET: Regions/Create
-        public IActionResult Create()
-        {
-            return View();
-        }
+        //// GET: Regions/Create
+        //public IActionResult Create()
+        //{
+        //    return View();
+        //}
 
-        // POST: Regions/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("Name,Location_Id,Id")] Region region)
-        {
-            if (ModelState.IsValid)
-            {
-                region.Id = Guid.NewGuid();
-                _context.Add(region);
-                await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
-            }
-            return View(region);
-        }
+        //// POST: Regions/Create
+        //// To protect from overposting attacks, enable the specific properties you want to bind to.
+        //// For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
+        //[HttpPost]
+        //[ValidateAntiForgeryToken]
+        //public async Task<IActionResult> Create([Bind("Name,Location_Id,Id")] Region region)
+        //{
+        //    if (ModelState.IsValid)
+        //    {
+        //        region.Id = Guid.NewGuid();
+        //        _context.Add(region);
+        //        await _context.SaveChangesAsync();
+        //        return RedirectToAction(nameof(Index));
+        //    }
+        //    return View(region);
+        //}
 
-        // GET: Regions/Edit/5
-        public async Task<IActionResult> Edit(Guid? id)
-        {
-            if (id == null || _context.Region == null)
-            {
-                return NotFound();
-            }
+        //// GET: Regions/Edit/5
+        //public async Task<IActionResult> Edit(Guid? id)
+        //{
+        //    if (id == null || _context.Region == null)
+        //    {
+        //        return NotFound();
+        //    }
 
-            var region = await _context.Region.FindAsync(id);
-            if (region == null)
-            {
-                return NotFound();
-            }
-            return View(region);
-        }
+        //    var region = await _context.Region.FindAsync(id);
+        //    if (region == null)
+        //    {
+        //        return NotFound();
+        //    }
+        //    return View(region);
+        //}
 
-        // POST: Regions/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(Guid id, [Bind("Name,Location_Id,Id")] Region region)
-        {
-            if (id != region.Id)
-            {
-                return NotFound();
-            }
+        //// POST: Regions/Edit/5
+        //// To protect from overposting attacks, enable the specific properties you want to bind to.
+        //// For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
+        //[HttpPost]
+        //[ValidateAntiForgeryToken]
+        //public async Task<IActionResult> Edit(Guid id, [Bind("Name,Location_Id,Id")] Region region)
+        //{
+        //    if (id != region.Id)
+        //    {
+        //        return NotFound();
+        //    }
 
-            if (ModelState.IsValid)
-            {
-                try
-                {
-                    _context.Update(region);
-                    await _context.SaveChangesAsync();
-                }
-                catch (DbUpdateConcurrencyException)
-                {
-                    if (!RegionExists(region.Id))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
-                }
-                return RedirectToAction(nameof(Index));
-            }
-            return View(region);
-        }
+        //    if (ModelState.IsValid)
+        //    {
+        //        try
+        //        {
+        //            _context.Update(region);
+        //            await _context.SaveChangesAsync();
+        //        }
+        //        catch (DbUpdateConcurrencyException)
+        //        {
+        //            if (!RegionExists(region.Id))
+        //            {
+        //                return NotFound();
+        //            }
+        //            else
+        //            {
+        //                throw;
+        //            }
+        //        }
+        //        return RedirectToAction(nameof(Index));
+        //    }
+        //    return View(region);
+        //}
 
-        // GET: Regions/Delete/5
-        public async Task<IActionResult> Delete(Guid? id)
-        {
-            if (id == null || _context.Region == null)
-            {
-                return NotFound();
-            }
+        //// GET: Regions/Delete/5
+        //public async Task<IActionResult> Delete(Guid? id)
+        //{
+        //    if (id == null || _context.Region == null)
+        //    {
+        //        return NotFound();
+        //    }
 
-            var region = await _context.Region
-                .FirstOrDefaultAsync(m => m.Id == id);
-            if (region == null)
-            {
-                return NotFound();
-            }
+        //    var region = await _context.Region
+        //        .FirstOrDefaultAsync(m => m.Id == id);
+        //    if (region == null)
+        //    {
+        //        return NotFound();
+        //    }
 
-            return View(region);
-        }
+        //    return View(region);
+        //}
 
-        // POST: Regions/Delete/5
-        [HttpPost, ActionName("Delete")]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(Guid id)
-        {
-            if (_context.Region == null)
-            {
-                return Problem("Entity set 'BoardBaronsApiContext.Region'  is null.");
-            }
-            var region = await _context.Region.FindAsync(id);
-            if (region != null)
-            {
-                _context.Region.Remove(region);
-            }
+        //// POST: Regions/Delete/5
+        //[HttpPost, ActionName("Delete")]
+        //[ValidateAntiForgeryToken]
+        //public async Task<IActionResult> DeleteConfirmed(Guid id)
+        //{
+        //    if (_context.Region == null)
+        //    {
+        //        return Problem("Entity set 'BoardBaronsApiContext.Region'  is null.");
+        //    }
+        //    var region = await _context.Region.FindAsync(id);
+        //    if (region != null)
+        //    {
+        //        _context.Region.Remove(region);
+        //    }
             
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
-        }
+        //    await _context.SaveChangesAsync();
+        //    return RedirectToAction(nameof(Index));
+        //}
 
-        private bool RegionExists(Guid id)
-        {
-          return (_context.Region?.Any(e => e.Id == id)).GetValueOrDefault();
-        }
+        //private bool RegionExists(Guid id)
+        //{
+        //  return (_context.Region?.Any(e => e.Id == id)).GetValueOrDefault();
+        //}
     }
 }
