@@ -5,160 +5,160 @@ using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
-using BoardBarons.Api.Data;
+//using BoardBarons.Api.Data;
 using BoardBarons.Business.Entities;
 
 namespace BoardBarons.Api.Controllers
 {
     public class UsersController : Controller
     {
-        private readonly BoardBaronsApiContext _context;
+        //private readonly BoardBaronsApiContext _context;
 
-        public UsersController(BoardBaronsApiContext context)
-        {
-            _context = context;
-        }
+        //public UsersController(BoardBaronsApiContext context)
+        //{
+        //    _context = context;
+        //}
 
-        // GET: Users
-        public async Task<IActionResult> Index()
-        {
-              return _context.User != null ? 
-                          View(await _context.User.ToListAsync()) :
-                          Problem("Entity set 'BoardBaronsApiContext.User'  is null.");
-        }
+        //// GET: Users
+        //public async Task<IActionResult> Index()
+        //{
+        //      return _context.User != null ? 
+        //                  View(await _context.User.ToListAsync()) :
+        //                  Problem("Entity set 'BoardBaronsApiContext.User'  is null.");
+        //}
 
-        // GET: Users/Details/5
-        public async Task<IActionResult> Details(Guid? id)
-        {
-            if (id == null || _context.User == null)
-            {
-                return NotFound();
-            }
+        //// GET: Users/Details/5
+        //public async Task<IActionResult> Details(Guid? id)
+        //{
+        //    if (id == null || _context.User == null)
+        //    {
+        //        return NotFound();
+        //    }
 
-            var user = await _context.User
-                .FirstOrDefaultAsync(m => m.Id == id);
-            if (user == null)
-            {
-                return NotFound();
-            }
+        //    var user = await _context.User
+        //        .FirstOrDefaultAsync(m => m.Id == id);
+        //    if (user == null)
+        //    {
+        //        return NotFound();
+        //    }
 
-            return View(user);
-        }
+        //    return View(user);
+        //}
 
-        // GET: Users/Create
-        public IActionResult Create()
-        {
-            return View();
-        }
+        //// GET: Users/Create
+        //public IActionResult Create()
+        //{
+        //    return View();
+        //}
 
-        // POST: Users/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("UserName,NormalizedUserName,Email,NormalizedEmail,EmailConfirmed,PasswordHash,SecurityStamp,ConcurrencyStamp,PhoneNumber,PhoneNumberConfirmed,Created_At,Desativated_At,Updated_At,TwoFactorEnabled,LockoutEnd,LockoutEnabled,AccessFailedCount,Region_Id,Id")] User user)
-        {
-            if (ModelState.IsValid)
-            {
-                user.Id = Guid.NewGuid();
-                _context.Add(user);
-                await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
-            }
-            return View(user);
-        }
+        //// POST: Users/Create
+        //// To protect from overposting attacks, enable the specific properties you want to bind to.
+        //// For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
+        //[HttpPost]
+        //[ValidateAntiForgeryToken]
+        //public async Task<IActionResult> Create([Bind("UserName,NormalizedUserName,Email,NormalizedEmail,EmailConfirmed,PasswordHash,SecurityStamp,ConcurrencyStamp,PhoneNumber,PhoneNumberConfirmed,Created_At,Desativated_At,Updated_At,TwoFactorEnabled,LockoutEnd,LockoutEnabled,AccessFailedCount,Region_Id,Id")] User user)
+        //{
+        //    if (ModelState.IsValid)
+        //    {
+        //        user.Id = Guid.NewGuid();
+        //        _context.Add(user);
+        //        await _context.SaveChangesAsync();
+        //        return RedirectToAction(nameof(Index));
+        //    }
+        //    return View(user);
+        //}
 
-        // GET: Users/Edit/5
-        public async Task<IActionResult> Edit(Guid? id)
-        {
-            if (id == null || _context.User == null)
-            {
-                return NotFound();
-            }
+        //// GET: Users/Edit/5
+        //public async Task<IActionResult> Edit(Guid? id)
+        //{
+        //    if (id == null || _context.User == null)
+        //    {
+        //        return NotFound();
+        //    }
 
-            var user = await _context.User.FindAsync(id);
-            if (user == null)
-            {
-                return NotFound();
-            }
-            return View(user);
-        }
+        //    var user = await _context.User.FindAsync(id);
+        //    if (user == null)
+        //    {
+        //        return NotFound();
+        //    }
+        //    return View(user);
+        //}
 
-        // POST: Users/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(Guid id, [Bind("UserName,NormalizedUserName,Email,NormalizedEmail,EmailConfirmed,PasswordHash,SecurityStamp,ConcurrencyStamp,PhoneNumber,PhoneNumberConfirmed,Created_At,Desativated_At,Updated_At,TwoFactorEnabled,LockoutEnd,LockoutEnabled,AccessFailedCount,Region_Id,Id")] User user)
-        {
-            if (id != user.Id)
-            {
-                return NotFound();
-            }
+        //// POST: Users/Edit/5
+        //// To protect from overposting attacks, enable the specific properties you want to bind to.
+        //// For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
+        //[HttpPost]
+        //[ValidateAntiForgeryToken]
+        //public async Task<IActionResult> Edit(Guid id, [Bind("UserName,NormalizedUserName,Email,NormalizedEmail,EmailConfirmed,PasswordHash,SecurityStamp,ConcurrencyStamp,PhoneNumber,PhoneNumberConfirmed,Created_At,Desativated_At,Updated_At,TwoFactorEnabled,LockoutEnd,LockoutEnabled,AccessFailedCount,Region_Id,Id")] User user)
+        //{
+        //    if (id != user.Id)
+        //    {
+        //        return NotFound();
+        //    }
 
-            if (ModelState.IsValid)
-            {
-                try
-                {
-                    _context.Update(user);
-                    await _context.SaveChangesAsync();
-                }
-                catch (DbUpdateConcurrencyException)
-                {
-                    if (!UserExists(user.Id))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
-                }
-                return RedirectToAction(nameof(Index));
-            }
-            return View(user);
-        }
+        //    if (ModelState.IsValid)
+        //    {
+        //        try
+        //        {
+        //            _context.Update(user);
+        //            await _context.SaveChangesAsync();
+        //        }
+        //        catch (DbUpdateConcurrencyException)
+        //        {
+        //            if (!UserExists(user.Id))
+        //            {
+        //                return NotFound();
+        //            }
+        //            else
+        //            {
+        //                throw;
+        //            }
+        //        }
+        //        return RedirectToAction(nameof(Index));
+        //    }
+        //    return View(user);
+        //}
 
-        // GET: Users/Delete/5
-        public async Task<IActionResult> Delete(Guid? id)
-        {
-            if (id == null || _context.User == null)
-            {
-                return NotFound();
-            }
+        //// GET: Users/Delete/5
+        //public async Task<IActionResult> Delete(Guid? id)
+        //{
+        //    if (id == null || _context.User == null)
+        //    {
+        //        return NotFound();
+        //    }
 
-            var user = await _context.User
-                .FirstOrDefaultAsync(m => m.Id == id);
-            if (user == null)
-            {
-                return NotFound();
-            }
+        //    var user = await _context.User
+        //        .FirstOrDefaultAsync(m => m.Id == id);
+        //    if (user == null)
+        //    {
+        //        return NotFound();
+        //    }
 
-            return View(user);
-        }
+        //    return View(user);
+        //}
 
-        // POST: Users/Delete/5
-        [HttpPost, ActionName("Delete")]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(Guid id)
-        {
-            if (_context.User == null)
-            {
-                return Problem("Entity set 'BoardBaronsApiContext.User'  is null.");
-            }
-            var user = await _context.User.FindAsync(id);
-            if (user != null)
-            {
-                _context.User.Remove(user);
-            }
+        //// POST: Users/Delete/5
+        //[HttpPost, ActionName("Delete")]
+        //[ValidateAntiForgeryToken]
+        //public async Task<IActionResult> DeleteConfirmed(Guid id)
+        //{
+        //    if (_context.User == null)
+        //    {
+        //        return Problem("Entity set 'BoardBaronsApiContext.User'  is null.");
+        //    }
+        //    var user = await _context.User.FindAsync(id);
+        //    if (user != null)
+        //    {
+        //        _context.User.Remove(user);
+        //    }
             
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
-        }
+        //    await _context.SaveChangesAsync();
+        //    return RedirectToAction(nameof(Index));
+        //}
 
-        private bool UserExists(Guid id)
-        {
-          return (_context.User?.Any(e => e.Id == id)).GetValueOrDefault();
-        }
+        //private bool UserExists(Guid id)
+        //{
+        //  return (_context.User?.Any(e => e.Id == id)).GetValueOrDefault();
+        //}
     }
 }
